separated Gamebox component into DailyGame and Unlimited to be better able to handle states for query
